Cleaned up extra component folder and renamed Register.js to index.js and fixed import
