+++ client/src/pages/RegisterPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Register from '../components/Register/Register'
+import Register from '../components/Register/'
 import API from "../utils/API"
 
 class RegisterPage extends Component {
